add sidebar menu , config and pssword rests

+++ app/Http/Controllers/Backend/ChargeCodeController.php
@@ -17,7 +17,7 @@ class ChargeCodeController extends Controller
     use DropdownTrait;
     public function index(Request $request): View|string
     {
-        //$this->authorize('viewAny', ChargeCode::class);
+        $this->authorize('viewAny', ChargeCode::class);
         $chargecodes = ChargeCode::companyOnly()->with('chargeGroup')->latest()->get();
 
         if ($request->ajax()) {
@@ -28,16 +28,16 @@ class ChargeCodeController extends Controller
     }
     public function create()
     {
-        //$this->authorize('create', ChargeCode::class);
+        $this->authorize('create', ChargeCode::class);
         return view('chargecodes.create', [
-            'insurances' => Insurance::all(),
+            'insurances' => Insurance::companyOnly()->get(),
             'groupTypes' => $this->getDropdownOptions('CHARGE_GROUP_TYPE')
         ]);
     }
 
     public function store(Request $request) : JsonResponse
     {
-        //$this->authorize('create', ChargeCode::class);
+        $this->authorize('create', ChargeCode::class);
         $request->validate([
             'chargeGroupType' => 'required|exists:drop_down_values,id',
             'code' => 'required|string|max:255|unique:charge_codes,code',
@@ -82,8 +82,8 @@ class ChargeCodeController extends Controller
      */
     public function show(ChargeCode $chargecode): View
     {
-        //$this->authorize('view', $chargecode);
-        $insurances = Insurance::all();
+        $this->authorize('view', $chargecode);
+        $insurances = Insurance::companyOnly()->get();
         $groupTypes = $this->getDropdownOptions('CHARGE_GROUP_TYPE');
 
         return view('chargecodes.show',compact('chargecode','insurances','groupTypes'));
@@ -91,15 +91,15 @@ class ChargeCodeController extends Controller
 
     public function edit(ChargeCode $chargecode): View
     {
-        //$this->authorize('update', $chargecode);
-        $insurances = Insurance::all();
+        $this->authorize('update', $chargecode);
+        $insurances = Insurance::companyOnly()->get();
         $groupTypes = $this->getDropdownOptions('CHARGE_GROUP_TYPE');
         return view('chargecodes.edit',compact('chargecode', 'insurances', 'groupTypes'));
     }
 
     public function update(Request $request, ChargeCode $chargecode): JsonResponse
     {
-        //$this->authorize('update', $chargecode);
+        $this->authorize('update', $chargecode);
         $request->validate([
             'code' => 'required|string|max:255',
             'chargeGroupType' => 'required|exists:drop_down_values,id',
@@ -142,7 +142,7 @@ class ChargeCodeController extends Controller
 
     public function destroy(ChargeCode $chargecode): RedirectResponse
     {
-        //$this->authorize('delete', $chargecode);
+        $this->authorize('delete', $chargecode);
         $chargecode->delete();
     
         return redirect()->route('chargecodes.index')
